+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   [#824](https://github.com/opendatateam/udata/issues/824)
 - Restore the broken URL check feature
   [#840](https://github.com/opendatateam/udata/issues/840)
+- Temporarily ignore INSPIRE in ODS harvester
+  [#837](https://github.com/opendatateam/udata/pull/837)
 
 ## 1.0.5 (2017-03-27)